Fixed a collision bug and changed player and enemy movement from move_and_slide to move_and_collide.
The bug occured when the player collided with the enemy but they were both travelling in the same direction (e.g. the enemy was heading up and the player was also heading up). This bug made it so that the enemy didn't register the collision so in effect, didn't kill the player.